factor improve

+++ src/main/java/com/mandarina/game/main/GameCts.java
@@ -3,7 +3,7 @@ package com.mandarina.game.main;
 public class GameCts {
 	public final static int EMPTY_TILE_VALUE = 128;
 	public final static int TILES_DEFAULT_SIZE = 32;
-	public final static double SCALE_DEFAULT = 1f;
+	public final static float SCALE_DEFAULT = 1f;
 	public final static int TILES_IN_WIDTH = 26;
 	public final static int TILES_IN_HEIGHT = 14;
 	public final static double GAME_WIDTH_DEFAULT = GameCts.TILES_DEFAULT_SIZE * GameCts.TILES_IN_WIDTH;
